rename Chip => ComboboxChip
