+++ packages/combobox/src/ComboboxChip/index.ts
@@ -1,0 +1,1 @@
+export { ComboboxChip } from './ComboboxChip';
