+++ packages/combobox/src/ComboboxOption/ComboboxOption.styles.ts
@@ -4,7 +4,7 @@ import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 import { fontWeights, spacing } from '@leafygreen-ui/tokens';
 
-import { fontSize, lineHeight } from '../Chip/Chip.styles';
+import { fontSize, lineHeight } from '../ComboboxChip/ComboboxChip.styles';
 import { ComboboxSize } from '../Combobox.types';
 import {
   getMenuItemHeight,
